+++ database/store/interface.go
