+++ database/store/leveldb.go
